Introduce `ActionView::Helpers::NavigationHelper`
Related to [#50312][]

Create the `ActionView::Helper::NavigationHelper` to serve as a module
for `button_to` and `link_to` (as well as related
`link_to_unless_current`, `link_to_unless`, and `link_to_if`).

As a side-effect of moving those methods, this commit also moves several
`private`-scoped methods related to those implementations. Since all
`ActionView::Helpers` modules are ultimately mixed-into the
`ActionView::Base` class, the `private` method boundaries are largely
ignored. Depsite that fact, this commit moves the methods so that they
are defined in the same module that they are called.

[#50312]: https://github.com/rails/rails/pull/50312#discussion_r2220564548
