+++ actionview/lib/action_view/helpers.rb
@@ -22,6 +22,7 @@ require "action_view/helpers/form_tag_helper"
 require "action_view/helpers/form_helper"
 require "action_view/helpers/form_options_helper"
 require "action_view/helpers/javascript_helper"
+require "action_view/helpers/navigation_helper"
 require "action_view/helpers/number_helper"
 require "action_view/helpers/rendering_helper"
 require "action_view/helpers/translation_helper"
@@ -56,6 +57,7 @@ module ActionView # :nodoc:
     include FormOptionsHelper
     include FormTagHelper
     include JavaScriptHelper
+    include NavigationHelper
     include NumberHelper
     include OutputSafetyHelper
     include RenderingHelper
